+++ Changelog.md
@@ -1,5 +1,21 @@
 # Changelog for MetroRadiance.Fork
 
+## 3.2.0 (preview) MetroRadiance
+
+- ### Breaking changes
+  - none
+
+- ### Enhancements/Features
+  - Add standard control styles (ListBox/ListView)
+  - Add attached properties (ListViewProperties)
+
+- ### Bug fixes
+  - none
+
+- ### Limitations
+  No changes
+  - When moving windows using AcrylicBlurWindow class, the move process is very slow on Windows 10 1903 (Build 18362, May 2019 Update) or later
+
 ## 3.1.1 MetroRadiance
 
 - ### Breaking changes